Add GB1 IgG binding landscape
Implements a new landscape on top of the Protein G GB1 domain fitness
data from Wu et al. This landscape is complete for the GB1 domain and
therefore operates as a lookup table, similar to the TF binding one.

This PR also includes the data on which this landscape is built.